+++ src/main/java/module-info.java
@@ -23,6 +23,7 @@
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 module comparator.extensions.main {
+	requires org.apache.commons.collections4;
 	exports io.github.astrapi69.comparator;
 	exports io.github.astrapi69.comparator.factory;
 	exports io.github.astrapi69.comparator.object;